requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
pypdf2
python-dotenv
ddgs

beautifulsoup4
requests
crewai
langgraph
reportlab
langsmith 
pdfplumber
pytesseract
Pillow
pymupdf
langchain-groq
groq